Fix PR Comments And WASM Comments
